Use bearer tokens for authentication.
Instead of passing our authentication token in the "auth_token" GET
parameter, we are now using the "Authorization" HTTP header as defined
in RFC 6750.

The advantage of this is that we don't have to worry about the URL
encoding of our Base64-encoded token.  Besides, middleboxes frequently
like to log request parameters (but not necessarily HTTP headers) and we
don't want our authentication token to be part of these logs.

Thanks to Arturo Filastò for suggesting this!